+++ src/main/scala/stores/store/valueobjects/ShelfOps.scala
@@ -9,9 +9,21 @@ package stores.store.valueobjects
 
 trait ShelfOps[A <: Shelf] {
 
-  def addItemsRow(itemsRow: ItemsRow): A
+  def addItemsRow(shelf: A, itemsRow: ItemsRow): A
 
-  def removeItemsRow(itemsRowId: ItemsRowId): A
+  def removeItemsRow(shelf: A, itemsRowId: ItemsRowId): A
 
-  def updateItemsRow(itemsRow: ItemsRow): A
+  def updateItemsRow(shelf: A, itemsRow: ItemsRow): A
+}
+
+object ShelfOps {
+
+  extension [A <: Shelf: ShelfOps](shelf: A) {
+
+    def addItemsRow(itemsRow: ItemsRow): A = implicitly[ShelfOps[A]].addItemsRow(shelf, itemsRow)
+
+    def removeItemsRow(itemsRowId: ItemsRowId): A = implicitly[ShelfOps[A]].removeItemsRow(shelf, itemsRowId)
+
+    def updateItemsRow(itemsRow: ItemsRow): A = implicitly[ShelfOps[A]].updateItemsRow(shelf, itemsRow)
+  }
 }