+++ src/main/scala/stores/store/valueobjects/ShelvingGroupOps.scala
@@ -9,9 +9,23 @@ package stores.store.valueobjects
 
 trait ShelvingGroupOps[A <: ShelvingGroup] {
 
-  def addShelving(shelving: Shelving): A
+  def addShelving(shelvingGroup: A, shelving: Shelving): A
 
-  def removeShelving(shelvingId: ShelvingId): A
+  def removeShelving(shelvingGroup: A, shelvingId: ShelvingId): A
 
-  def updateShelving(shelving: Shelving): A
+  def updateShelving(shelvingGroup: A, shelving: Shelving): A
+}
+
+object ShelvingGroupOps {
+
+  extension [A <: ShelvingGroup: ShelvingGroupOps](shelvingGroup: A) {
+
+    def addShelving(shelving: Shelving): A = implicitly[ShelvingGroupOps[A]].addShelving(shelvingGroup, shelving)
+
+    def removeShelving(shelvingId: ShelvingId): A =
+      implicitly[ShelvingGroupOps[A]].removeShelving(shelvingGroup, shelvingId)
+
+    def updateShelving(shelving: Shelving): A =
+      implicitly[ShelvingGroupOps[A]].updateShelving(shelvingGroup, shelving)
+  }
 }