+++ src/main/scala/stores/store/valueobjects/ShelvingOps.scala
@@ -9,9 +9,21 @@ package stores.store.valueobjects
 
 trait ShelvingOps[A <: Shelving] {
 
-  def addShelf(shelf: Shelf): A
+  def addShelf(shelving: A, shelf: Shelf): A
 
-  def removeShelf(shelfId: ShelfId): A
+  def removeShelf(shelving: A, shelfId: ShelfId): A
 
-  def updateShelf(shelf: Shelf): A
+  def updateShelf(shelving: A, shelf: Shelf): A
+}
+
+object ShelvingOps {
+
+  extension [A <: Shelving: ShelvingOps](shelving: A) {
+
+    def addShelf(shelf: Shelf): A = implicitly[ShelvingOps[A]].addShelf(shelving, shelf)
+
+    def removeShelf(shelfId: ShelfId): A = implicitly[ShelvingOps[A]].removeShelf(shelving, shelfId)
+
+    def updateShelf(shelf: Shelf): A = implicitly[ShelvingOps[A]].updateShelf(shelving, shelf)
+  }
 }